Sort files & renames part 2
