+++ src/components/LandingComponents/Hero/AnimateSideHero.component.js
@@ -1,9 +1,9 @@
 'use client'
-import { ImageDemoLeftComponent } from '@/components/ImageDemoComponents/ImageDemoLeft.component'
-import { ImageDemoMidTopComponent } from '@/components/ImageDemoComponents/ImageDemoMidTop.component'
-import { ImageDemoMidBottomComponent } from '@/components/ImageDemoComponents/ImageDemoMidBottom.component'
-import { ImageDemoRightTopComponent } from '@/components/ImageDemoComponents/ImageDemoRightTop.component'
-import { ImageDemoRightBottomComponent } from '@/components/ImageDemoComponents/ImageDemoRightBottom.component'
+import { ImageDemoLeftComponent } from '@/components/LandingComponents/Hero/ImageDemoComponents/ImageDemoLeft.component'
+import { ImageDemoMidTopComponent } from '@/components/LandingComponents/Hero/ImageDemoComponents/ImageDemoMidTop.component'
+import { ImageDemoMidBottomComponent } from '@/components/LandingComponents/Hero/ImageDemoComponents/ImageDemoMidBottom.component'
+import { ImageDemoRightTopComponent } from '@/components/LandingComponents/Hero/ImageDemoComponents/ImageDemoRightTop.component'
+import { ImageDemoRightBottomComponent } from '@/components/LandingComponents/Hero/ImageDemoComponents/ImageDemoRightBottom.component'
 import Image from 'next/image'
 import { useEffect, useState } from 'react'
